Added puppeteerProduct to check Puppeteer version, moved logStyle to switch-browser.js, and added self-invoking main in backstopjs-test.js
